refactor: use PackageInstaller where applicable
